requirements.txt
torch
torchvision
Pillow
numpy
gradio